backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pillow==10.2.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
drf-yasg==1.21.7
setuptools>=65.5.0
pytz==2024.1
packaging>=23.1